fix: Trainloading type and dependent code
