+++ app/graphql/types/tfl/entities/crowding_type.rb
@@ -3,7 +3,7 @@ module Types
     module Entities
       class CrowdingType < Types::BaseObject
         field :passengerFlows, [Types::Tfl::Entities::PassengerFlowType, null: true], null: true
-        field :trainLoadings, [Types::TrainLoadingType, null: true], null: true
+        field :trainLoadings, [Types::Tfl::Entities::TrainLoadingType, null: true], null: true
       end
     end
   end
